+++ changelog.md
@@ -12,11 +12,15 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Added
 
--
+- Add Media Admin
+- Add Media Admin Page in admin
+- Add List and delete images in Storage and open containing documents (events or news)
+- Add Migration page and functions for images and image related databases
+- Add services for loading and deleting images and getting an image's size
 
 ### Changed
 
--
+- Metadata management for event and news editor: read from database not from metadata
 
 ### Fixed